backend/requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
pymongo==4.7.2
bcrypt==4.1.3
python-dotenv==1.0.1
google-generativeai==0.7.2
groq==0.9.0
